auxiliary WCS values
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@10441 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ lib/rts2/connection.cpp
@@ -2106,7 +2106,7 @@ double Connection::getProgress (double now)
 
 DataAbstractRead* Connection::lastDataChannel (int chan)
 {
-	if (readChannels.size () == 0 || (--readChannels.end ())->second->size () <= chan)
+	if (readChannels.size () == 0 || ((int) (--readChannels.end ())->second->size ()) <= chan)
 		return NULL;
 	return (--readChannels.end ())->second->at (chan);
 }
